More Overhual to Player Class/Creator
There is a bunch of notes that you can find in mostly the Player.cs file, and some in the Enums.cs file. Have put very rough systems with baseline stats and a player creator with preset classes/stats already chosen for the player, in the future we may let the player change these stat values themselves. Also added a stat page for gameplay, still gotta push for an easier way for the player to pull the stats up.